requirements.txt
setuptools==75.6.0
#numpy==2.1.3
#numpy-stl==3.2.0
paho-mqtt==2.1.0
python-telegram-bot==21.8
standard-imghdr==3.13.0
urllib3==2.2.3
python-dateutil>=2.9.0
matplotlib==3.9.3
